DRY ArticleKeywordExtractor
